Consolidate all google service account authentication
